Minor refactoring of the user role system
Rename the usermnagement package to users, add a few comments
