+++ app/src/main/java/ch/epfl/sweng/eventmanager/ui/eventSelector/EventPickingActivity.java
@@ -16,12 +16,12 @@ import android.widget.TextView;
 import ch.epfl.sweng.eventmanager.R;
 import ch.epfl.sweng.eventmanager.ui.userManager.DisplayAccountActivity;
 import ch.epfl.sweng.eventmanager.ui.userManager.LoginActivity;
-import ch.epfl.sweng.eventmanager.userManagement.Session;
+import ch.epfl.sweng.eventmanager.users.Session;
 import ch.epfl.sweng.eventmanager.viewmodel.ViewModelFactory;
 import dagger.android.AndroidInjection;
 
 import javax.inject.Inject;
-import java.util.ArrayList;
+
 import java.util.Collections;
 
 public class EventPickingActivity extends AppCompatActivity {
