--- app/src/main/java/ch/epfl/sweng/eventmanager/userManagement/Role.java
@@ -1,6 +1,0 @@
-package ch.epfl.sweng.eventmanager.userManagement;
-
-public enum Role {
-    ADMIN,
-    STAFF
-}
