+++ app/src/main/java/ch/epfl/sweng/eventmanager/users/Role.java
@@ -1,0 +1,9 @@
+package ch.epfl.sweng.eventmanager.users;
+
+/**
+ * Defines the clearance levels for authenticated users accessing an event.
+ */
+public enum Role {
+    ADMIN,
+    STAFF
+}
