--- app/src/test/java/ch/epfl/sweng/eventmanager/userManagement/SessionTest.java
@@ -1,12 +1,0 @@
-package ch.epfl.sweng.eventmanager.userManagement;
-
-import ch.epfl.sweng.eventmanager.repository.data.User;
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-/**
- * @author Louis Vialar
- */
-public class SessionTest {
-}
