+++ app/src/test/java/ch/epfl/sweng/eventmanager/users/SessionTest.java
@@ -1,0 +1,7 @@
+package ch.epfl.sweng.eventmanager.users;
+
+/**
+ * @author Louis Vialar
+ */
+public class SessionTest {
+}
